Switch over to using the Anomaly Likelihood class in NuPIC
Changes:

Switched over to using the Anomaly Likelihood class.   Initially there was a performance drop due to some small differences in how the distribution estimation is handled.  I fixed these in NuPIC.  With those changes we get identical performance to what we see without using this class.

I also retuned the parameters a bit and improved overall scores.

I've verified that switching to log likelihood does not affect the final score, so we now output that as our anomaly score. It is much easier to plot. Interestingly the best threshold is consistently 0.5.

Fixes #172