Update ARTime README.md
Details in the text but the basic problem is that ARTime does not run out of the box now because of updates to the Julia language. The user will need to install Julia 1.7.0 If there is demand I could look into updating the ARTime Julia code to support the latest Julia version but that is not high on my list right now.